Added start/stop/queue events for download manager.
