+++ src/events/connection.event.js
@@ -1,5 +1,5 @@
 const logger = require('../logger');
 
-module.exports = function connection(args) {
+module.exports = function connection() {
     logger.info(`Client connected. ID: ${this.client.id}`);
 };
