+++ src/events/default.event.js
@@ -4,6 +4,6 @@
 
 const logger = require('../logger');
 
-module.exports = function defaultHandler(eventName, model, data) {
+module.exports = function defaultHandler(eventName, mgr, data) {
     logger.warn(`Uncaught event [${eventName}] from client ${this.client.id}: ${(JSON).stringify(data)}`);
 };
