+++ src/events/mgr-start.event.js
@@ -1,0 +1,5 @@
+const Manager = require('../manager');
+
+module.exports = function mgrStart(mgr) {
+    Manager.run(mgr);
+};
