+++ src/events/mgr-stop.event.js
@@ -1,0 +1,5 @@
+const Manager = require('../manager');
+
+module.exports = function mgrStop(mgr) {
+    Manager.stop(mgr);
+};
